[modularize] Move remote function related functions into `Distributor`.
